Optimize data& add module 'api'

+++ gradf.framework.data/src/main/java/liangchen/wang/gradf/framework/data/dao/ISequenceDao.java
@@ -5,5 +5,5 @@ package liangchen.wang.gradf.framework.data.dao;
  */
 public interface ISequenceDao {
 
-    Long sequenceNumber(String sequenceKey, int initialValue);
+    Long sequenceNumber(String sequenceKey, long initialValue);
 }
