+++ gradf.framework.data/src/main/java/liangchen/wang/gradf/framework/data/manager/ISequenceManager.java
@@ -4,7 +4,7 @@ package liangchen.wang.gradf.framework.data.manager;
  * @author LiangChen.Wang
  */
 public interface ISequenceManager {
-    Long sequenceNumber(String sequenceKey, int initialValue);
+    Long sequenceNumber(String sequenceKey, long initialValue);
 
     Long sequenceNumber(String sequenceKey);
 }
